fix(form): set captcha token on submit instead of on fetch
The token is only valid once. If it is set directly after retrieving it, it'll be used for the
first request following and unset directly afterwards. In case of the registration form
that would be likely the username existence check and the token would not be valid
for the form submission later on.
Now we set the token on form submission only.